+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # GoSysMess Changelog
 Все изменения библиотеки GoSysMess будут документироваться на этой странице.
 
-## 2023-09-02
-### Add
+## 2023-09-03
+### Changed
+- ErrorId -> string
+- FieldErrorList -> *FieldErrorList